Modularize meeting detection lifecycle
Separate listener context wiring, meeting app presentation, auto-stop policy, and native detect-event orchestration behind the existing public context API without changing behavior.
